made functions system in generator-elements -> issues
