+++ src/styles.css
@@ -2,4 +2,5 @@ body{
     background-color: black;
     color: dimgrey;
     margin: 0;
+    overflow: hidden;
 }
